# perf: bench async op baseline