This extends the `op_baseline` bench introduced in https://github.com/denoland/deno/pull/9924 to also bench async ops.

Here's a comparison of `main` vs https://github.com/denoland/deno/pull/9947:

```
Before:
test bench_op_async   ... bench:   1,256,422 ns/iter (+/- 27,839)
test bench_op_nop     ... bench:     323,275 ns/iter (+/- 7,916)
test bench_op_pi_bin  ... bench:     810,688 ns/iter (+/- 26,193)
test bench_op_pi_json ... bench:     703,911 ns/iter (+/- 44,627)

After:
test bench_op_async   ... bench:     805,011 ns/iter (+/- 22,248)
test bench_op_nop     ... bench:     318,809 ns/iter (+/- 11,228)
test bench_op_pi_bin  ... bench:     408,557 ns/iter (+/- 21,002)
test bench_op_pi_json ... bench:     315,826 ns/iter (+/- 23,131)
```